move Attribute, When, WhenOperator to annotation/attributes.py
